requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
amqp==5.3.1
    # via
    #   -r requirements/base.txt
    #   kombu
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
    #   django-countries
billiard==4.2.1
    # via
    #   -r requirements/base.txt
    #   celery
celery==5.4.0
    # via -r requirements/base.txt
certifi==2025.1.31
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.17.1
    # via
    #   -r requirements/base.txt
    #   cryptography
    #   pynacl
chardet==5.2.0
    # via diff-cover
charset-normalizer==3.4.1
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.8
    # via
    #   -r requirements/base.txt
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   code-annotations
    #   edx-django-utils
click-didyoumean==0.3.1
    # via
    #   -r requirements/base.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r requirements/base.txt
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements/base.txt
    #   celery
code-annotations==2.2.0
    # via -r requirements/test.in
coverage[toml]==7.6.12
    # via pytest-cov
cryptography==44.0.2
    # via
    #   -r requirements/base.txt
    #   django-fernet-fields-v2
    #   jwcrypto
    #   pyjwt
ddt==1.3.1
    # via -r requirements/test.in
diff-cover==9.2.3
    # via -r requirements/test.in
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-config-models
    #   django-crum
    #   django-fernet-fields-v2
    #   django-model-utils
    #   django-oauth-toolkit
    #   django-waffle
    #   djangorestframework
    #   drf-jwt
    #   edx-django-utils
    #   edx-drf-extensions
    #   jsonfield
django-config-models==2.8.0
    # via -r requirements/base.txt
django-countries==7.6.1
    # via -r requirements/test.in
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
django-fernet-fields-v2==0.9
    # via -r requirements/base.txt
django-model-utils==5.0.0
    # via
    #   -r requirements/base.txt
    #   -r requirements/test.in
django-oauth-toolkit==3.0.1
    # via -r requirements/base.txt
django-object-actions==4.3.0
    # via -r requirements/base.txt
django-simple-history==3.0.0
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
django-waffle==4.2.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.15.2
    # via
    #   -r requirements/base.txt
    #   django-config-models
    #   drf-jwt
    #   edx-drf-extensions
dnspython==2.7.0
    # via
    #   -r requirements/base.txt
    #   pymongo
drf-jwt==1.19.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-django-utils==7.2.0
    # via
    #   -r requirements/base.txt
    #   django-config-models
    #   edx-drf-extensions
edx-drf-extensions==10.5.0
    # via -r requirements/base.txt
edx-opaque-keys[django]==2.11.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
factory-boy==3.3.3
    # via -r requirements/test.in
faker==36.2.2
    # via factory-boy
freezegun==1.5.1
    # via -r requirements/test.in
idna==3.10
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.6
    # via
    #   code-annotations
    #   diff-cover
jsonfield==3.1.0
    # via -r requirements/base.txt
jwcrypto==1.5.6
    # via
    #   -r requirements/base.txt
    #   django-oauth-toolkit
kombu==5.4.2
    # via
    #   -r requirements/base.txt
    #   celery
markupsafe==3.0.2
    # via jinja2
mock==5.2.0
    # via -r requirements/test.in
newrelic==10.7.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
oauthlib==3.2.2
    # via
    #   -r requirements/base.txt
    #   django-oauth-toolkit
openedx-atlas==0.6.2
    # via -r requirements/base.txt
packaging==24.2
    # via pytest
pbr==6.1.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.5.0
    # via
    #   diff-cover
    #   pytest
prompt-toolkit==3.0.50
    # via
    #   -r requirements/base.txt
    #   click-repl
psutil==7.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.22
    # via
    #   -r requirements/base.txt
    #   cffi
pygments==2.19.1
    # via diff-cover
pyjwt[crypto]==2.10.1
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
pymongo==4.11.2
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==8.3.5
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==6.0.0
    # via -r requirements/test.in
pytest-django==4.10.0
    # via -r requirements/test.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/base.txt
    #   celery
    #   freezegun
python-slugify==8.0.4
    # via code-annotations
pytz==2025.1
    # via -r requirements/base.txt
pyyaml==6.0.2
    # via
    #   code-annotations
    #   responses
requests==2.32.3
    # via
    #   -r requirements/base.txt
    #   django-oauth-toolkit
    #   edx-drf-extensions
    #   responses
    #   slumber
responses==0.25.6
    # via -r requirements/test.in
semantic-version==2.10.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
six==1.17.0
    # via
    #   -r requirements/base.txt
    #   python-dateutil
slumber==0.7.1
    # via -r requirements/base.txt
sqlparse==0.5.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.4.1
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
testfixtures==8.3.0
    # via -r requirements/test.in
text-unidecode==1.3
    # via python-slugify
typing-extensions==4.12.2
    # via
    #   -r requirements/base.txt
    #   django-countries
    #   edx-opaque-keys
    #   jwcrypto
tzdata==2025.1
    # via
    #   -r requirements/base.txt
    #   celery
    #   faker
    #   kombu
unicodecsv==0.14.1
    # via -r requirements/base.txt
urllib3==2.2.3
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   requests
    #   responses
vine==5.1.0
    # via
    #   -r requirements/base.txt
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via
    #   -r requirements/base.txt
    #   prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
